Refactor all the models

+++ api/livematch/management/commands/prunematches.py
@@ -2,11 +2,10 @@ import logging
 from datetime import timedelta
 from django.conf import settings
 from django.core.management.base import BaseCommand
-from django.db.models import Q
+from django.db.models import Q, Max
 from django.utils import timezone
-from itertools import chain
 
-from livematch.models import LiveMatch, LivePlayerMatch
+from livematch.models import LiveMatch
 
 
 logger = logging.getLogger(settings.RPS_LOGGER_NAME)
@@ -22,24 +21,29 @@ class Command(BaseCommand):
             default=60 * 60 * 24,  # 24h
             help="How long since a match was active (in seconds) to prune it",
         )
+        parser.add_argument(
+            "--dry-run", action="store_true", help="Don't actuall delete shit"
+        )
 
-    def handle(self, *args, ttl, **options):
+    def handle(self, *args, ttl, dry_run, **options):
         logger.info("Pruning LiveMatch...")
         cutoff_time = timezone.now() - timedelta(seconds=ttl)
-        matches_to_delete = LiveMatch.objects.filter(
-            Q(player1=None) | Q(player1__last_activity__lt=cutoff_time),
-            Q(player2=None) | Q(player2__last_activity__lt=cutoff_time),
+        matches_to_delete = LiveMatch.objects.annotate(
+            npa=Max("liveplayermatch__last_activity")
+        ).filter(
+            Q(npa=None) | Q(npa__lt=cutoff_time),
             start_time__lt=cutoff_time,
+            # We can't delete matches that have been rematched, because they
+            # are still needed to link child Match -> parent Match
+            # The leaf child will be deleted once it ages out, and that will
+            # cascade up the chain
+            rematch_id=None,
         )
-        # LivePlayerMatch is a target of an FK on LiveMatch, so they don't
-        # cascade automatically - we have to gather and delete them manually
-        pm_ids_to_delete = set(
-            chain(*matches_to_delete.values_list("player1_id", "player2_id"))
-        )
-        total, per_model = matches_to_delete.delete()
-        pm_total, pm_per_model = LivePlayerMatch.objects.filter(
-            id__in=pm_ids_to_delete
-        ).delete()
-        total += pm_total
-        per_model.update(pm_per_model)
-        logger.info(f"Deleted {total} rows ({per_model})")
+
+        if dry_run:
+            logger.info(
+                f"Skipping deleting {matches_to_delete.count()} matches"
+            )
+        else:
+            total, per_model = matches_to_delete.delete()
+            logger.info(f"Deleted {total} rows ({per_model})")
